Movendo de static para ssr
