+++ src/pages/_document.js
@@ -1,0 +1,38 @@
+/* eslint-disable react/no-danger, import/no-unresolved, jsx-a11y/html-has-lang */
+import React from 'react';
+import Document, { Head, Main, NextScript } from 'next/document';
+import { extractStyles } from 'evergreen-ui';
+import flush from 'styled-jsx/server';
+
+export default class MyDocument extends Document {
+  static getInitialProps({ renderPage }) {
+    const page = renderPage();
+    const styles = flush();
+    const { css, hydrationScript } = extractStyles();
+
+    return {
+      ...page,
+      css,
+      hydrationScript,
+      styles
+    };
+  }
+
+  render() {
+    const { css, hydrationScript } = this.props;
+
+    return (
+      <html>
+        <Head>
+          <style dangerouslySetInnerHTML={{ __html: css }} />
+        </Head>
+
+        <body>
+          <Main />
+          {hydrationScript}
+          <NextScript />
+        </body>
+      </html>
+    );
+  }
+}
