+++ src/pages/dashboard.js
@@ -1,0 +1,5 @@
+import Page from '../layouts/page';
+
+const Index = () => <Page title="Dashboard">dashboard</Page>;
+
+export default Index;
